@@ -60,71 +28,94 @@
   "author": "prescottprue (https://github.com/prescottprue)",
   "license": "MIT",
   "dependencies": {
-    "firebase": "^4.1.3",
-    "lodash": "^4.17.2",
-    "material-ui": "0.16.0",
-    "normalize.css": "^4.1.1",
+    "lodash": "^4.17.4",
+    "material-ui": "^0.18.6",
+    "normalize.css": "^7.0.0",
+    "object-assign": "^4.1.1",
+    "promise": "^7.1.1",
     "prop-types": "^15.5.10",
-    "react": "15.3.2",
-    "react-dom": "15.3.2",
-    "react-google-button": "^0.1.0",
-    "react-redux": "^4.4.5",
-    "react-redux-firebase": "2.0.0-alpha.7",
-    "react-router": "^2.8.0",
-    "react-tap-event-plugin": "1.0.0",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-google-button": "^0.3.0",
+    "react-redux": "^5.0.4",
+    "react-redux-firebase": "canary",
+    "react-router": "^3.0.0",
+    "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.6.0",
     "redux-auth-wrapper": "^1.0.0",
     "redux-form": "^6.6.1",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0"
+    "redux-form-material-ui": "^4.2.0",
+    "redux-thunk": "^2.2.0",
+    "redbox-react": "^1.3.6",
+    "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
-    "babel-core": "^6.22.1",
-    "babel-eslint": "^6.0.4",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-istanbul": "^2.0.1",
-    "babel-plugin-lodash": "^3.2.10",
+    "babel-core": "^6.24.1",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.0.0",
+    "babel-plugin-lodash": "^3.2.11",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-polyfill": "^6.16.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-0": "^6.3.13",
-    "babel-runtime": "^6.11.6",
-    "better-npm-run": "^0.0.13",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-runtime": "^6.20.0",
+    "chai": "^3.5.0",
+    "chai-as-promised": "^6.0.0",
+    "chai-enzyme": "^0.6.1",
+    "chalk": "^1.1.3",
+    "codecov": "^2.2.0",
+    "compression": "^1.6.2",
     "connect-history-api-fallback": "^1.3.0",
-    "css-loader": "^0.25.0",
-    "cssnano": "^3.7.4",
-    "debug": "^2.2.0",
-    "eslint": "^3.9.1",
+    "cross-env": "^5.0.0",
+    "css-loader": "^0.28.1",
+    "dirty-chai": "^1.2.2",
+    "enzyme": "^2.8.2",
+    "eslint": "^3.19.0",
     "eslint-config-prettier": "^2.3.0",
-    "eslint-config-standard": "^6.2.1",
-    "eslint-config-standard-react": "^4.2.0",
-    "eslint-plugin-babel": "^3.3.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-config-standard-react": "^5.0.0",
+    "eslint-plugin-babel": "^4.1.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-node": "^4.2.2",
+    "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-prettier": "^2.1.2",
-    "eslint-plugin-promise": "^3.3.1",
-    "eslint-plugin-react": "^6.0.0",
-    "eslint-plugin-standard": "^2.0.0",
+    "eslint-plugin-react": "^7.0.1",
+    "eslint-plugin-standard": "^3.0.1",
     "express": "^4.14.0",
-    "extract-text-webpack-plugin": "^1.0.0",
-    "file-loader": "^0.9.0",
-    "fs-extra": "^1.0.0",
-    "html-webpack-plugin": "^2.22.0",
-    "imports-loader": "^0.6.5",
-    "ip": "^1.1.2",
-    "json-loader": "^0.5.4",
-    "node-sass": "^3.7.0",
-    "nodemon": "^1.10.2",
-    "postcss-loader": "^0.13.0",
-    "prettier": "^1.5.2",
-    "redbox-react": "^1.2.10",
-    "rimraf": "^2.5.4",
-    "sass-loader": "^4.0.0",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.12.14",
-    "webpack-dev-middleware": "1.10.1",
-    "webpack-hot-middleware": "^2.12.2",
-    "yargs": "^6.3.0"
+    "extract-text-webpack-plugin": "^2.1.0",
+    "favicons-webpack-plugin": "0.0.7",
+    "figures": "^2.0.0",
+    "file-loader": "^0.11.1",
+    "fs-extra": "^3.0.1",
+    "html-webpack-plugin": "^2.29.0",
+    "husky": "^0.14.3",
+    "ip": "^1.1.5",
+    "istanbul-instrumenter-loader": "^3.0.0-rc.1",
+    "karma": "^1.7.0",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
+    "karma-mocha": "^1.3.0",
+    "karma-mocha-reporter": "^2.2.1",
+    "karma-phantomjs-launcher": "^1.0.4",
+    "karma-webpack-with-fast-source-maps": "^1.10.0",
+    "mocha": "^3.2.0",
+    "node-sass": "^4.5.3",
+    "phantomjs-prebuilt": "^2.1.14",
+    "prettier": "^1.5.3",
+    "react-addons-test-utils": "^15.5.1",
+    "react-test-renderer": "^15.5.4",
+    "rimraf": "^2.6.1",
+    "sass-loader": "^6.0.5",
+    "sinon": "^2.2.0",
+    "sinon-chai": "^2.10.0",
+    "style-loader": "^0.17.0",
+    "url-loader": "^0.5.8",
+    "webpack": "^2.5.1",
+    "webpack-dashboard": "^0.4.0",
+    "webpack-dev-middleware": "^1.11.0",
+    "webpack-hot-middleware": "^2.13.2",
+    "yargs": "^8.0.1"
   }
 }
