@@ -1,15 +1,13 @@
 const express = require('express')
-const debug = require('debug')('app:server')
+const path = require('path')
 const webpack = require('webpack')
-const webpackConfig = require('../config/webpack.config')
-const project = require('../config/project.config')
+const logger = require('../build/lib/logger')
+const webpackConfig = require('../build/webpack.config')
+const project = require('../project.config')
+const compress = require('compression')
 
 const app = express()
-
-// This rewrites all routes requests to the root /index.html file
-// (ignoring file requests). If you want to implement universal
-// rendering, you'll want to remove this middleware.
-app.use(require('connect-history-api-fallback')())
+app.use(compress())
 
 // ------------------------------------
 // Apply Webpack HMR Middleware
@@ -17,36 +15,60 @@ app.use(require('connect-history-api-fallback')())
 if (project.env === 'development') {
   const compiler = webpack(webpackConfig)
 
-  debug('Enable webpack dev and HMR middleware')
-  app.use(require('webpack-dev-middleware')(compiler, {
-    publicPath: webpackConfig.output.publicPath,
-    contentBase: project.paths.client(),
-    hot: true,
-    quiet: project.compiler_quiet,
-    noInfo: project.compiler_quiet,
-    lazy: false,
-    stats: project.compiler_stats
-  }))
-  app.use(require('webpack-hot-middleware')(compiler))
+  logger.info('Enabling webpack development and HMR middleware')
+  app.use(
+    require('webpack-dev-middleware')(compiler, {
+      publicPath: webpackConfig.output.publicPath,
+      headers: {
+        'Access-Control-Allow-Origin': '*'
+      },
+      contentBase: path.resolve(project.basePath, project.srcDir),
+      hot: true,
+      quiet: false,
+      noInfo: false,
+      lazy: false,
+      stats: 'normal'
+    })
+  )
+  app.use(
+    require('webpack-hot-middleware')(compiler, {
+      path: '/__webpack_hmr'
+    })
+  )
 
-  // Serve static assets from ~/src/static since Webpack is unaware of
+  // Serve static assets from ~/public since Webpack is unaware of
   // these files. This middleware doesn't need to be enabled outside
   // of development since this directory will be copied into ~/dist
   // when the application is compiled.
-  app.use(express.static(project.paths.client('static')))
-} else {
-  debug(
-    'Server is being run outside of live development mode, meaning it will ' +
-    'only serve the compiled application bundle in ~/dist. Generally you ' +
-    'do not need an application server for this and can instead use a web ' +
-    'server such as nginx to serve your static files. See the "deployment" ' +
-    'section in the README for more information on deployment strategies.'
-  )
+  app.use(express.static(path.resolve(project.basePath, 'public')))
 
+  // This rewrites all routes requests to the root /index.html file
+  // (ignoring file requests). If you want to implement universal
+  // rendering, you'll want to remove this middleware.
+  app.use('*', function(req, res, next) {
+    const filename = path.join(compiler.outputPath, 'index.html')
+    compiler.outputFileSystem.readFile(filename, (err, result) => {
+      if (err) {
+        return next(err)
+      }
+      res.set('content-type', 'text/html')
+      res.send(result)
+      res.end()
+    })
+  })
+} else {
+  logger.warn(
+    'Server is being run outside of live development mode, meaning it will ' +
+      'only serve the compiled application bundle in ~/dist. Generally you ' +
+      'do not need an application server for this and can instead use a web ' +
+      'server such as nginx to serve your static files. See the "deployment" ' +
+      'section in the README for more information on deployment strategies.'
+  )
+  // NOTE: If you are using this, you should make express and compress dependencies instead of dev dependencies
   // Serving ~/dist by default. Ideally these files should be served by
   // the web server and not the app server, but this helps to demo the
   // server in production.
-  app.use(express.static(project.paths.dist()))
+  app.use(express.static(path.resolve(project.basePath, project.outDir)))
 }
 
 module.exports = app
