@@ -6,7 +6,7 @@ import classes from './LoadingSpinner.scss'
 export const LoadingSpinner = ({ size }) =>
   <div className={classes.container}>
     <div className={classes.progress}>
-      <CircularProgress mode="indeterminate" size={size || 80} />
+      <CircularProgress mode='indeterminate' size={size || 80} />
     </div>
   </div>
 
