@@ -2,6 +2,8 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { browserHistory, Router } from 'react-router'
 import { Provider } from 'react-redux'
+
+// Themeing/Styling
 import Theme from 'theme'
 import getMuiTheme from 'material-ui/styles/getMuiTheme'
 
@@ -14,24 +16,22 @@ export default class AppContainer extends Component {
     muiTheme: PropTypes.object
   }
 
-  getChildContext = () => ({
-    muiTheme: getMuiTheme(Theme)
-  })
-
   static propTypes = {
     routes: PropTypes.object.isRequired,
     store: PropTypes.object.isRequired
   }
 
+  getChildContext = () => ({
+    muiTheme: getMuiTheme(Theme)
+  })
+
   render() {
     const { routes, store } = this.props
     return (
       <Provider store={store}>
-        <div style={{ height: '100%' }}>
-          <Router history={browserHistory}>
-            {routes}
-          </Router>
-        </div>
+        <Router history={browserHistory}>
+          {routes}
+        </Router>
       </Provider>
     )
   }
