@@ -1,17 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import classes from './Navbar.scss'
 import { Link } from 'react-router'
-import { connect } from 'react-redux'
-import {
-  firebaseConnect,
-  pathToJS,
-  isLoaded,
-  isEmpty
-} from 'react-redux-firebase'
-import { LIST_PATH, ACCOUNT_PATH, LOGIN_PATH, SIGNUP_PATH } from 'constants'
-
-// Components
 import AppBar from 'material-ui/AppBar'
 import IconMenu from 'material-ui/IconMenu'
 import IconButton from 'material-ui/IconButton'
@@ -19,7 +8,15 @@ import MenuItem from 'material-ui/MenuItem'
 import FlatButton from 'material-ui/FlatButton'
 import DownArrow from 'material-ui/svg-icons/hardware/keyboard-arrow-down'
 import Avatar from 'material-ui/Avatar'
+import { connect } from 'react-redux'
+import {
+  firebaseConnect,
+  isLoaded,
+  isEmpty
+} from 'react-redux-firebase'
+import { LIST_PATH, ACCOUNT_PATH, LOGIN_PATH, SIGNUP_PATH } from 'constants'
 import defaultUserImage from 'static/User.png'
+import classes from './Navbar.scss'
 
 const buttonStyle = {
   color: 'white',
@@ -52,6 +49,7 @@ export default class Navbar extends Component {
 
   static propTypes = {
     profile: PropTypes.object,
+    auth: PropTypes.object,
     firebase: PropTypes.object.isRequired
   }
 
@@ -60,66 +58,67 @@ export default class Navbar extends Component {
     this.context.router.push('/')
   }
 
-  render () {
+  render() {
     const { profile, auth } = this.props
-    const profileExists = isLoaded(profile) && !isEmpty(profile)
+    const dataLoaded = isLoaded(auth, profile)
     const authExists = isLoaded(auth) && !isEmpty(auth)
 
     const iconButton = (
       <IconButton style={avatarStyles.button} disableTouchRipple>
         <div className={classes.avatar}>
-          <div className="hidden-mobile">
+          <div className='hidden-mobile'>
             <Avatar
-              src={profileExists && profile.avatarUrl ? profile.avatarUrl : defaultUserImage}
+              src={
+                profile && profile.avatarUrl
+                  ? profile.avatarUrl
+                  : defaultUserImage
+              }
             />
           </div>
           <div className={classes['avatar-text']}>
             <span className={`${classes['avatar-text-name']} hidden-mobile`}>
-              { profileExists && profile.displayName ? profile.displayName : 'User' }
+              {profile && profile.displayName ? profile.displayName : 'User'}
             </span>
-            <DownArrow color="white" />
+            <DownArrow color='white' />
           </div>
         </div>
       </IconButton>
     )
 
-    const mainMenu = (
-      <div className={classes.menu}>
-        <Link to={SIGNUP_PATH}>
-          <FlatButton label="Sign Up" style={buttonStyle} />
-        </Link>
-        <Link to={LOGIN_PATH}>
-          <FlatButton label="Login" style={buttonStyle} />
-        </Link>
-      </div>
-    )
-
-    const rightMenu = authExists ? (
-      <IconMenu
-        iconButtonElement={iconButton}
-        targetOrigin={{ horizontal: 'right', vertical: 'bottom' }}
-        anchorOrigin={{ horizontal: 'right', vertical: 'top' }}
-        animated={false}
-      >
-        <MenuItem
-          primaryText='Account'
-          onTouchTap={() => this.context.router.push(ACCOUNT_PATH)}
-        />
-        <MenuItem
-          primaryText='Sign out'
-          onTouchTap={this.handleLogout}
-        />
-      </IconMenu>
-    ) : mainMenu
+    const rightMenu = dataLoaded && authExists
+      ? (
+        <IconMenu
+          iconButtonElement={iconButton}
+          targetOrigin={{ horizontal: 'right', vertical: 'bottom' }}
+          anchorOrigin={{ horizontal: 'right', vertical: 'top' }}
+          animated={false}
+        >
+          <MenuItem
+            primaryText='Account'
+            onTouchTap={() => this.context.router.push(ACCOUNT_PATH)}
+          />
+          <MenuItem
+            primaryText='Sign out'
+            onTouchTap={this.handleLogout}
+          />
+        </IconMenu>
+      )
+      : (
+        <div className={classes.menu}>
+          <Link to={SIGNUP_PATH}>
+            <FlatButton label='Sign Up' style={buttonStyle} />
+          </Link>
+          <Link to={LOGIN_PATH}>
+            <FlatButton label='Login' style={buttonStyle} />
+          </Link>
+        </div>
+      )
 
     return (
       <AppBar
         title={
-          <Link
-            to={authExists ? `${LIST_PATH}` : '/'}
-            className={classes.brand}
-          >
-            material example
+          <Link to={authExists ? LIST_PATH : '/'} className={classes.brand}>
+            material
           </Link>
         }
         showMenuIconButton={false}
