@@ -2,22 +2,14 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import Paper from 'material-ui/Paper'
 import { isObject } from 'lodash'
-import IconButton from 'material-ui/IconButton'
-import DeleteIcon from 'material-ui/svg-icons/action/delete'
-
 import classes from './ProjectTile.scss'
 
-export const ProjectTile = ({ project, onSelect, onDelete, showDelete }) =>
+export const ProjectTile = ({ project, onSelect }) =>
   <Paper className={classes.container}>
     <div className={classes.top}>
       <span className={classes.name} onClick={() => onSelect(project)}>
         {project.name}
       </span>
-      {showDelete && onDelete
-        ? <IconButton tooltip="delete" onClick={onDelete}>
-            <DeleteIcon />
-          </IconButton>
-        : null}
     </div>
     <span className={classes.owner}>
       {isObject(project.createdBy)
@@ -28,9 +20,7 @@ export const ProjectTile = ({ project, onSelect, onDelete, showDelete }) =>
 
 ProjectTile.propTypes = {
   project: PropTypes.object.isRequired,
-  onSelect: PropTypes.func.isRequired,
-  onDelete: PropTypes.func.isRequired,
-  showDelete: PropTypes.bool
+  onSelect: PropTypes.func.isRequired
 }
 
 export default ProjectTile
