@@ -6,6 +6,7 @@ import SignupRoute from './Signup'
 import ProjectsRoute from './Projects'
 import AccountRoute from './Account'
 import RecoverRoute from './Recover'
+import NotFoundRoute from './NotFound'
 
 /*  Note: Instead of using JSX, we recommend using react-router
     PlainRoute objects to build route definitions.   */
@@ -19,7 +20,9 @@ export const createRoutes = store => ({
     LoginRoute,
     SignupRoute,
     ProjectsRoute(store), // async route definitions recieve store
-    RecoverRoute(store) // async route definitions recieve store
+    RecoverRoute(store), // async route definitions recieve store
+    /* Place all Routes above here so NotFoundRoute can act as a 404 page */
+    NotFoundRoute(store)
   ]
 })
 
